# docs/openapi/with_auth.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cohere Toolkit API",
    "description": "This is the API for the Open Source Cohere Toolkit",
    "version": "v1.1.6"
  },
  "paths": {
...
    "/v1/users": {
      "post": {
        "tags": ["user"],
        "summary": "Create User",
        "description": "Create a new user.",
        "operationId": "create_user_v1_users_post",
        "parameters": [
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/backend__schemas__user__CreateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/backend__schemas__user__User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["user"],
        "summary": "List Users",
        "description": "List all users.",
        "operationId": "list_users_v1_users_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Pagination Offset",
              "description": "Offset for where request should start returning records from",
              "default": 0
            },
            "description": "Offset for where request should start returning records from"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Pagination Limit",
              "description": "Maximum number of records to return per request",
              "default": 100
            },
            "description": "Maximum number of records to return per request"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/backend__schemas__user__User"
                  },
                  "title": "Response List Users V1 Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "tags": ["user"],
        "summary": "Get User",
        "description": "Get a user by ID.\n\nRaises:\n    HTTPException: If the user with the given ID is not found.",
        "operationId": "get_user_v1_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User ID",
              "description": "User ID for the user in question"
            },
            "description": "User ID for the user in question"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/backend__schemas__user__User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["user"],
        "summary": "Update User",
        "description": "Update a user by ID.\n\nRaises:\n    HTTPException: If the user with the given ID is not found.",
        "operationId": "update_user_v1_users__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User ID",
              "description": "User ID for the user in question"
            },
            "description": "User ID for the user in question"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/backend__schemas__user__UpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/backend__schemas__user__User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "summary": "Delete User",
        "description": "Delete a user by ID.\n\nRaises:\n    HTTPException: If the user with the given ID is not found.",
        "operationId": "delete_user_v1_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User ID",
              "description": "User ID for the user in question"
            },
            "description": "User ID for the user in question"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteUser" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/v1/organizations/{org_id}/users": {
      "get": {
        "tags": ["organization"],
        "summary": "Get Organization Users",
        "description": "Get organization users by ID.",
        "operationId": "get_organization_users_v1_organizations__org_id__users_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization ID",
              "description": "Organization ID for the organization in question"
            },
            "description": "Organization ID for the organization in question"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization",
              "description": "Authorization header containing Bearer token"
            },
            "description": "Authorization header containing Bearer token"
          },
          {
            "name": "Organization-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Organization ID",
              "description": "Unique Identifier for the Organization making the request"
            },
            "description": "Unique Identifier for the Organization making the request"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/backend__schemas__user__User"
                  },
                  "title": "Response Get Organization Users V1 Organizations  Org Id  Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/scim/v2/Users": {
      "get": {
        "tags": ["scim"],
        "summary": "Get Users",
        "description": "Return users",
        "operationId": "get_users_scim_v2_Users_get",
        "security": [{ "HTTPBasic": [] }],
        "parameters": [
          {
            "name": "start_index",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Start Index",
              "description": "Start Index for request",
              "default": 1
            },
            "description": "Start Index for request"
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Count",
              "description": "Maximum number of records to return per request",
              "default": 100
            },
            "description": "Maximum number of records to return per request"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Filter",
              "description": "Filter to use when filtering response"
            },
            "description": "Filter to use when filtering response"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListUserResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["scim"],
        "summary": "Create User",
        "description": "Create a new user",
        "operationId": "create_user_scim_v2_Users_post",
        "security": [{ "HTTPBasic": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/backend__schemas__scim__CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/scim/v2/Users/{user_id}": {
      "get": {
        "tags": ["scim"],
        "summary": "Get User",
        "description": "Get user by User ID",
        "operationId": "get_user_scim_v2_Users__user_id__get",
        "security": [{ "HTTPBasic": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User ID",
              "description": "User ID for the user in question"
            },
            "description": "User ID for the user in question"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["scim"],
        "summary": "Update User",
        "description": "Update a user",
        "operationId": "update_user_scim_v2_Users__user_id__put",
        "security": [{ "HTTPBasic": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User ID",
              "description": "User ID for the user in question"
            },
            "description": "User ID for the user in question"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/backend__schemas__scim__UpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["scim"],
        "summary": "Patch User",
        "description": "Patch a user",
        "operationId": "patch_user_scim_v2_Users__user_id__patch",
        "security": [{ "HTTPBasic": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User ID",
              "description": "User ID for the user in question"
            },
            "description": "User ID for the user in question"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchUser" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteUser": {
        "properties": {},
        "type": "object",
        "title": "DeleteUser",
        "description": "Response when deleting a user"
      },
...
      "Email": {
        "properties": {
          "primary": {
            "type": "boolean",
            "title": "Primary",
            "description": "Is email the primary email"
          },
          "value": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Value",
            "description": "Email value"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Type of email"
          }
        },
        "type": "object",
        "required": ["primary", "type"],
        "title": "Email"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListUserResponse": {
        "properties": {
          "totalResults": {
            "type": "integer",
            "title": "Total Results",
            "description": "Total results available"
          },
          "startIndex": {
            "type": "integer",
            "title": "Start Index",
            "description": "Start index for returned results"
          },
          "itemsPerPage": {
            "type": "integer",
            "title": "Items Per Page",
            "description": "Total results returned in the request"
          },
          "Resources": {
            "items": {
              "$ref": "#/components/schemas/backend__schemas__scim__User"
            },
            "type": "array",
            "title": "Resources",
            "description": "List of Users"
          }
        },
        "type": "object",
        "required": ["totalResults", "startIndex", "itemsPerPage", "Resources"],
        "title": "ListUserResponse"
      },
...
      "Meta": {
        "properties": {
          "resourceType": {
            "type": "string",
            "title": "Resource Type",
            "description": "Type of resource the metadata is for"
          },
          "created": {
            "type": "string",
            "title": "Created",
            "description": "When metadata was created"
          },
          "lastModified": {
            "type": "string",
            "title": "Last Modified",
            "description": "When metadata was last modified"
          }
        },
        "type": "object",
        "required": ["resourceType", "created", "lastModified"],
        "title": "Meta",
        "description": "Schema for metadata"
      },
...
      "Name": {
        "properties": {
          "givenName": {
            "type": "string",
            "title": "Given Name",
            "description": "User's given name"
          },
          "familyName": {
            "type": "string",
            "title": "Family Name",
            "description": "User's family name"
          }
        },
        "type": "object",
        "required": ["givenName", "familyName"],
        "title": "Name"
      },
...
      "Operation": {
        "properties": {
          "op": { "type": "string", "title": "Op", "description": "Op" },
          "value": {
            "additionalProperties": { "type": "boolean" },
            "type": "object",
            "title": "Value",
            "description": "Value"
          }
        },
        "type": "object",
        "required": ["op", "value"],
        "title": "Operation"
      },
...
      "PatchUser": {
        "properties": {
          "schemas": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Schemas",
            "description": "Schemas for user"
          },
          "operations": {
            "items": { "$ref": "#/components/schemas/Operation" },
            "type": "array",
            "title": "Operations",
            "description": "Operations for the user"
          }
        },
        "type": "object",
        "required": ["schemas", "operations"],
        "title": "PatchUser"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "backend__schemas__scim__CreateUser": {
        "properties": {
          "userName": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Name",
            "description": "User name"
          },
          "active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Active",
            "description": "Is user active"
          },
          "schemas": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Schemas",
            "description": "Schemas for the user"
          },
          "name": {
            "$ref": "#/components/schemas/Name",
            "title": "Name",
            "description": "Name of user"
          },
          "emails": {
            "items": { "$ref": "#/components/schemas/Email" },
            "type": "array",
            "title": "Emails",
            "description": "List of emails for user"
          },
          "externalId": {
            "type": "string",
            "title": "External ID",
            "description": "External ID for the user"
          }
        },
        "type": "object",
        "required": ["schemas", "name", "emails", "externalId"],
        "title": "CreateUser"
      },
      "backend__schemas__scim__UpdateUser": {
        "properties": {
          "userName": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Name",
            "description": "User name"
          },
          "active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Active",
            "description": "Is user active"
          },
          "schemas": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Schemas",
            "description": "Schemas for the user"
          },
          "name": {
            "$ref": "#/components/schemas/Name",
            "description": "Name of user"
          },
          "emails": {
            "items": { "$ref": "#/components/schemas/Email" },
            "type": "array",
            "title": "Emails",
            "description": "List of emails for user"
          }
        },
        "type": "object",
        "required": ["schemas", "name", "emails"],
        "title": "UpdateUser"
      },
      "backend__schemas__scim__User": {
        "properties": {
          "userName": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Name",
            "description": "User name"
          },
          "active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Active",
            "description": "Is user active"
          },
          "schemas": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Schemas",
            "description": "Schemas for the user"
          },
          "id": {
            "type": "string",
            "title": "ID",
            "description": "Unique identifier for the user"
          },
          "externalId": {
            "type": "string",
            "title": "External ID",
            "description": "External ID for the user"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta",
            "description": "Metadata for the user"
          }
        },
        "type": "object",
        "required": ["schemas", "id", "externalId", "meta"],
        "title": "User"
      },
      "backend__schemas__user__CreateUser": {
        "properties": {
          "password": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Password",
            "description": "Password for the user"
          },
          "hashed_password": {
            "anyOf": [
              { "type": "string", "format": "binary" },
              { "type": "null" }
            ],
            "title": "Hashed Password",
            "description": "The user's password hashed"
          },
          "fullname": {
            "type": "string",
            "title": "Full Name",
            "description": "User's Full Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email",
            "description": "User's email address"
          }
        },
        "type": "object",
        "required": ["fullname"],
        "title": "CreateUser",
        "description": "Request to create a user"
      },
      "backend__schemas__user__UpdateUser": {
        "properties": {
          "password": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Password",
            "description": "Password for the user"
          },
          "hashed_password": {
            "anyOf": [
              { "type": "string", "format": "binary" },
              { "type": "null" }
            ],
            "title": "Hashed Password",
            "description": "The user's password hashed"
          },
          "fullname": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Full Name",
            "description": "User's Full Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email",
            "description": "User's email address"
          }
        },
        "type": "object",
        "title": "UpdateUser",
        "description": "Request to update a user"
      },
      "backend__schemas__user__User": {
        "properties": {
          "fullname": {
            "type": "string",
            "title": "Full Name",
            "description": "User's Full Name"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email",
            "description": "User's email address"
          },
          "id": { "type": "string", "title": "ID", "description": "" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Timestamp",
            "description": "When the user was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At Timestamp",
            "description": "When the user was updated"
          }
        },
        "type": "object",
        "required": ["fullname", "id", "created_at", "updated_at"],
        "title": "User",
        "description": "User schema"
      }
    },
    "securitySchemes": { "HTTPBasic": { "type": "http", "scheme": "basic" } }
  }
}